SoundUI 추가
SoundUI 추가
- SoundUI를 추가해 주었다.
- SoundUI에는 Sound의 이름과 경로가 뜨며, Sound를 재생 시켜볼 수 있고, Sound의 볼륨을 실시간으로 변경해 확인해볼 수 있다.

추가 사항
- UI의 Close함수를 ResUI에서 override해주었고, SoundUI에서도 override로 UI가 Close될 경우 사운드를 멈춰주도록 만들어 주었다.

경고 몇 가지 해결
- InspecterUI의 멤버 변수 초기화 추가 m_PprefabObject , m_pLevelPath nullptr로 초기화

+++ 01. Project/TemProgram/ListUI_EX.cpp
@@ -91,7 +91,7 @@ void ListUI_EX::render_update()
 
 				if (ImGui::Selectable(m_vecItemName[i].c_str(), Selectable))
 				{
-					m_iSelectIdx = i;
+					m_iSelectIdx = (UINT)i;
 				}
 
 				// show selected
